+++ pkgs/google_generative_ai/CHANGELOG.md
@@ -4,6 +4,8 @@
 - Add named constructors on `Schema` for each value type.
 - Add `GenerationConfig.responseMimeType` which supports setting
   `'application/json'` to force the model to reply with JSON parseable output.
+- Add `outputDimensionality` argument support for `embedContent` and
+  `batchEmbedContent`.
 
 ## 0.3.3